+++ Sources/SwifterSwift/SwiftStdlib/SequenceExtensions.swift
@@ -282,6 +282,16 @@ public extension Sequence {
     func first<T: Equatable>(where keyPath: KeyPath<Element, T>, equals value: T) -> Element? {
         return first { $0[keyPath: keyPath] == value }
     }
+
+    /// SwifterSwift: Returns the last element of the sequence with having property by given key path equals to given `value`.
+    ///
+    /// - Parameters:
+    ///   - keyPath: The `KeyPath` of property for `Element` to compare.
+    ///   - value: The value to compare with `Element` property
+    /// - Returns: The last element of the collection that has property by given key path equals to given `value` or `nil` if there is no such element.
+    func last<T: Equatable>(where keyPath: KeyPath<Element, T>, equals value: T) -> Element? {
+        return last { $0[keyPath: keyPath] == value }
+    }
 }
 
 public extension Sequence where Element: Equatable {